requirements.txt
attrs==19.1.0
Automat==0.7.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
defusedxml==0.6.0
Django==2.2.2
django-cors-middleware==1.5.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-photologue==3.10
django-sortedm2m==2.0.0
djangorestframework==3.10.3
ExifRead==2.1.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1.1
oauthlib==3.1.0
Pillow==6.2.0
pycparser==2.20
PyHamcrest==1.9.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.3.2
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.6.1
zope.interface==4.6.0
